requirements.txt
beautifulsoup4
selenium
python-dotenv
google-generativeai
python-dotenv 
pymongo
fastapi
uvicorn